src/train_stage1.py
"""
Stage 1 Training: Pre-training phase
"""

import argparse
import yaml
import torch
import torch.nn as nn
from torch.utils.data import DataLoader
import os


def train_stage1(config):
    """
    Stage 1 training logic
    """
    print("Starting Stage 1 training...")
    # TODO: Implement training logic
    pass


def main():
    parser = argparse.ArgumentParser(description='ACTFormer Stage 1 Training')
    parser.add_argument('--config', type=str, required=True,
                       help='Path to config file')
    parser.add_argument('--gpu', type=int, default=0,
                       help='GPU device ID')
    parser.add_argument('--seed', type=int, default=42,
                       help='Random seed')
    
    args = parser.parse_args()
    
    # Load configuration
    with open(args.config, 'r') as f:
        config = yaml.safe_load(f)
    
    # Set device
    device = torch.device(f'cuda:{args.gpu}' if torch.cuda.is_available() else 'cpu')
    print(f"Using device: {device}")
    
    # Set random seed
    torch.manual_seed(args.seed)
    
    # Train
    train_stage1(config)


if __name__ == '__main__':
    main()
