requirements.txt
# Core dependencies
torch>=1.10.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Deep learning utilities
pytorch-lightning>=1.5.0
tensorboard>=2.8.0

# Data processing
scikit-learn>=1.0.0
h5py>=3.6.0
pyyaml>=6.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Graph processing
torch-geometric>=2.0.0
networkx>=2.6.0

# Utilities
tqdm>=4.62.0
einops>=0.4.0
